requirements.txt
absl-py==2.4.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
astroid==4.0.3
attrs==25.4.0
autoflake==2.3.1
autopep8==2.3.2
black==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
datasets==4.5.0
dill==0.4.0
einops==0.8.2
evaluate==0.4.6
faiss-gpu-cu12==1.13.2
filelock==3.20.0
flash_attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.8.3/flash_attn-2.8.3+cu12torch2.8cxx11abiTRUE-cp311-cp311-linux_x86_64.whl#sha256=3d41b2fc55753faa7f45d6568ea73a96b96afb48b82994ab9b49bcbcb6c87588
frozenlist==1.8.0
fsspec==2025.10.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.11
isort==7.0.0
Jinja2==3.1.6
joblib==1.5.3
libcst==1.8.6
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.6.77
packaging @ file:///home/task_176104885106445/conda-bld/packaging_1761049078006/work
pandas==3.0.0
pathspec==1.0.4
peft==0.18.1
pillow==12.0.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.0
pycln==2.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
ruff==0.15.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
shap==0.50.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.8.0+cu126
torchaudio==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.3
transformers==5.1.0
triton==3.4.0
typer==0.21.1
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
xgboost==3.1.3
xxhash==3.6.0
yarl==1.22.0
